Added PerformanceAnalyser class
Used Apache Spark SizeEstimator to get a current size estimate of 32B for QTNode and 40B for Point objects.  

Estimates memory used based on these sizes and the number of each class that would need to be stored.

Measures time taken to compile both structures (QuadTree or TreeSet<Point<Integer>>).